refactor: move UserLoginRequestDto and UserRegisterRequestDto to dto.auth package
