+++ src/main/java/com/challengeteam/shop/dto/auth/UserRegisterRequestDto.java
@@ -1,4 +1,4 @@
-package com.challengeteam.shop.dto.user;
+package com.challengeteam.shop.dto.auth;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
