+++ src/test/java/com/challengeteam/shop/web/TestAuthHelper.java
@@ -1,8 +1,8 @@
 package com.challengeteam.shop.web;
 
 import com.challengeteam.shop.dto.jwt.JwtResponseDto;
-import com.challengeteam.shop.dto.user.UserLoginRequestDto;
-import com.challengeteam.shop.dto.user.UserRegisterRequestDto;
+import com.challengeteam.shop.dto.auth.UserLoginRequestDto;
+import com.challengeteam.shop.dto.auth.UserRegisterRequestDto;
 import com.challengeteam.shop.exceptionHandling.exception.EmailAlreadyExistsException;
 import com.challengeteam.shop.service.JwtAuthorizationService;
 import lombok.RequiredArgsConstructor;
